+++ libft/libft.h
@@ -37,7 +37,7 @@ typedef struct s_list
 } t_list;
 
 int		ft_atoi(const char *str);
-int     ft_atol(const char *str);
+long    ft_atol(const char *str);
 void	ft_bzero(void *dst, size_t size);
 void	*ft_calloc(size_t nmemb, size_t size);
 int		ft_isalnum(int c);
